worker/requirements.txt
mysql-connector-python==8.0.17
requests
pandas
gunicorn
pika
kazoo
